# lab3/src/main/resources/API.yaml
{
  "openapi": "3.0.1",
  "info": {
    "title": "Lab 2 Rest API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/company/rooms/{id}": {
      "get": {
        "tags": [
          "room-rest-controller"
        ],
        "summary": "Получить данные помещения",
        "description": "Получение данных помещения по его ID",
        "operationId": "read_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID помещения",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успешное выполнение",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RoomDTO"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RoomDTO"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "room-rest-controller"
        ],
        "summary": "Добавить/изменить помещение",
        "description": "Добавление/изменение помещения в зависимости от существования указанного ID",
        "operationId": "create",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID помещения",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "name",
            "in": "query",
            "description": "Название",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "department",
            "in": "query",
            "description": "Отдел",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "capacity",
            "in": "query",
            "description": "Вместительность",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успешное выполнение",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RoomDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "room-rest-controller"
        ],
        "summary": "Удалить помещение",
        "description": "Удаление помещения по ID",
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID помещения",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успешное выполнение",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/company/employees/{id}": {
      "get": {
        "tags": [
          "employee-rest-controller"
        ],
        "summary": "Получить данные сотрудника",
        "description": "Получение данных сотрудника по его ID",
        "operationId": "read_3",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID сотрудника",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успешное выполнение",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeDTO"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeDTO"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "employee-rest-controller"
        ],
        "summary": "Добавить/изменить сотрудника",
        "description": "Добавление/изменение сотрудника в зависимости от существования указанного ID",
        "operationId": "create_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID сотрудника",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "name",
            "in": "query",
            "description": "ФИО",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "department",
            "in": "query",
            "description": "Отдел",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "position",
            "in": "query",
            "description": "Должность",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успешное выполнение",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "employee-rest-controller"
        ],
        "summary": "Удалить сотрудника",
        "description": "Удаление сотрудника по ID",
        "operationId": "delete_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID сотрудника",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успешное выполнение",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/company/rooms": {
      "get": {
        "tags": [
          "room-rest-controller"
        ],
        "summary": "Получить список помещений",
        "description": "Получение всех помещений и их данных",
        "operationId": "read",
        "responses": {
          "200": {
            "description": "Успешное выполнение",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoomDTO"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoomDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/company/employees": {
      "get": {
        "tags": [
          "employee-rest-controller"
        ],
        "summary": "Получить список сотрудников",
        "description": "Получение всех сотрудников и их данных",
        "operationId": "read_2",
        "responses": {
          "200": {
            "description": "Успешное выполнение",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EmployeeDTO"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EmployeeDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "RoomDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID помещения",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "description": "Название помещения"
          },
          "department": {
            "type": "string",
            "description": "Отдел, которому принадлежит помещение"
          },
          "capacity": {
            "type": "integer",
            "description": "Вместительность помещения",
            "format": "int32"
          }
        },
        "description": "Помещение"
      },
      "EmployeeDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID сотрудника",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "description": "Имя сотрудника"
          },
          "department": {
            "type": "string",
            "description": "Отдел, где работает сотрудник"
          },
          "position": {
            "type": "string",
            "description": "Должность сотрудника"
          }
        },
        "description": "Сотрудник"
      },
...
    }
  }
}